--- a/app/seeders/basic_data/builtin_roles_seeder.rb
+++ b/app/seeders/basic_data/builtin_roles_seeder.rb
@@ -30,7 +30,7 @@ module BasicData
     def seed_data!
       data.each do |attributes|
         unless Role.find_by(builtin: attributes[:builtin]).nil?
-          puts "   *** Skipping built in role #{attributes[:name]} - already exists"
+          print_status "   *** Skipping built in role #{attributes[:name]} - already exists"
           next
         end