--- a/app/seeders/basic_data/type_seeder.rb
+++ b/app/seeders/basic_data/type_seeder.rb
@@ -72,25 +72,27 @@ module BasicData
     end
 
     def set_attribute_groups_for_type(type)
-      return if demo_data_for('type_configuration').nil?
+      type_data = type_data_for(type)
+      return unless type_data && type_data['form_configuration']
 
-      demo_data_for('type_configuration').each do |entry|
-        if entry[:form_configuration] && I18n.t(entry[:type]) === type.name
+      type_data['form_configuration'].each do |form_config_attr|
+        groups = type.default_attribute_groups
+        query = find_query_by_name(form_config_attr['query_name'])
+        query_association = "query_#{query}"
+        groups.unshift([form_config_attr['group_name'], [query_association.to_sym]])
 
-          entry[:form_configuration].each do |form_config_attr|
-            groups = type.default_attribute_groups
-            query_association = 'query_' + find_query_by_name(form_config_attr[:query_name]).to_s
-            groups.unshift([form_config_attr[:group_name], [query_association.to_sym]])
-
-            type.attribute_groups = groups
-          end
-
-          type.save!
-        end
+        type.attribute_groups = groups
       end
+
+      type.save!
     end
 
     private
+
+    def type_data_for(type)
+      types_data = seed_data.lookup('type_configuration') || []
+      types_data.find { |entry| I18n.t(entry['type']) == type.name }
+    end
 
     def find_query_by_name(name)
       Query.find_by(name:).id