--- a/app/seeders/composite_seeder.rb
+++ b/app/seeders/composite_seeder.rb
@@ -27,18 +27,19 @@
 class CompositeSeeder < Seeder
   def seed_data!
     ActiveRecord::Base.transaction do
-      data_seeders.each do |seeder|
-        puts " ↳ #{seeder.class.name.demodulize}"
-        seeder.seed!
-      end
+      seed_with(data_seeders)
 
-      return if discovered_seeders.empty?
-
-      puts "   Loading discovered seeders: "
-      discovered_seeders.each do |seeder|
-        puts " ↳ #{seeder.class.name.demodulize}"
-        seeder.seed!
+      if discovered_seeders.any?
+        print_status "Loading discovered seeders: #{discovered_seeders.map { seeder_name(_1) }.join(', ')}"
+        seed_with(discovered_seeders)
       end
+    end
+  end
+
+  def seed_with(seeders)
+    seeders.each do |seeder|
+      print_status " ↳ #{seeder_name(seeder)}"
+      seeder.seed!
     end
   end
 
@@ -75,5 +76,9 @@ class CompositeSeeder < Seeder
   # Accepts plugin seeders, e.g. 'BasicData::Documents'.
   def include_discovered_class?(discovered_class)
     discovered_class.name =~ /^#{namespace}::/
+  end
+
+  def seeder_name(seeder)
+    seeder.class.name.split('::').without(namespace).join('::')
   end
 end