--- a/app/seeders/demo_data/group_seeder.rb
+++ b/app/seeders/demo_data/group_seeder.rb
@@ -29,13 +29,6 @@ module DemoData
   class GroupSeeder < Seeder
     include ::DemoData::References
 
-    attr_reader :seed_data
-
-    def initialize(seed_data)
-      super()
-      @seed_data = seed_data
-    end
-
     def seed_data!
       print_status '    ↳ Creating groups' do
         seed_groups