--- a/app/seeders/demo_data/project_seeder.rb
+++ b/app/seeders/demo_data/project_seeder.rb
@@ -84,7 +84,6 @@ module DemoData
     def project_data_seeders(project, key)
       seeders = [
         DemoData::WikiSeeder,
-        DemoData::CustomFieldSeeder,
         DemoData::WorkPackageSeeder,
         DemoData::WorkPackageBoardSeeder
       ]