--- a/app/seeders/demo_data/project_seeder.rb
+++ b/app/seeders/demo_data/project_seeder.rb
@@ -70,9 +70,6 @@ module DemoData
         Setting.demo_projects_available = true
       end
 
-      puts ' ↳ Assign groups to projects'
-      set_groups
-
       puts ' ↳ Update form configuration with global queries'
       set_form_configuration
     end
@@ -147,10 +144,6 @@ module DemoData
         principal: user,
         roles: [role]
       )
-    end
-
-    def set_groups
-      DemoData::GroupSeeder.new.add_projects_to_groups
     end
 
     def set_form_configuration