completes https://community.openproject.org/wp/47352
part of https://community.openproject.org/wp/36933

- [x] Move seeding logic in one place
  - two places actually: `/app/seeders` and `/modules/bim/app/seeders`
- [x] Decouple seeding and translation file
- [x] Use dedicated logger to print to stdout when calling `rails db:seed`

As a side effect, seeding got slightly faster.
Unit test spec/seeders/demo_data_seeder_spec.rb
before: Finished in 8.51 seconds (files took 8.71 seconds to load)
after: Finished in 5.07 seconds (files took 9.18 seconds to load)